Is Fibo
You are given an integer, NN. Write a program to determine if NN is an
element of the Fibonacci sequence.

The first few elements of the Fibonacci sequence are
0,1,1,2,3,5,8,13,⋯0,1,1,2,3,5,8,13,⋯. A Fibonacci sequence is one where
every element is a sum of the previous two elements in the sequence.
The first two elements are 00 and 11.

Formally:
fib0fib1⋮fibn=0=1=fibn−1+fibn−2∀n>1
fib0=0fib1=1⋮fibn=fibn−1+fibn−2∀n>1

Input Format
The first line contains TT, number of test cases.
TT lines follow. Each line contains an integer NN.

Output Format
Display IsFibo if NN is a Fibonacci number and IsNotFibo if it is not.
The output for each test case should be displayed in a new line.

Constraints
1≤T≤1051≤T≤105
1≤N≤10101≤N≤1010

Sample Input

3
5
7
8

Sample Output

IsFibo
IsNotFibo
IsFibo

Explanation
55 is a Fibonacci number given by fib5=3+2fib5=3+2
77 is not a Fibonacci number
88 is a Fibonacci number given by fib6=5+3fib6=5+3

Time Limit
Time limit for this challenge is given here.